requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
termcolor==2.5.0
tenacity==9.0.0